+++ src/t/bootstrap/test/ptrs.cr
@@ -4,21 +4,21 @@ import helpers.log;
 import helpers.assert;
 
 export proc
-ptr_arith3_t_12345_r2(p: i32*, n: i64): i32
+ptr_arith3_t_12345_r2(p: i32*, n: size_t): i32
 {
         log::id("ptrs::ptr_arith3_t_12345_r2");
         return p[-1];
 }
 
 export proc
-ptr_arith2_t_12345_r2(p: i32*, n: i64): i32
+ptr_arith2_t_12345_r2(p: i32*, n: size_t): i32
 {
         log::id("ptrs::ptr_arith2_t_12345_r2");
         return *(p-1);
 }
 
 export proc
-ptr_arith1_t_12345_r2(p: i32*, n: i64): i32
+ptr_arith1_t_12345_r2(p: i32*, n: size_t): i32
 {
         log::id("ptrs::ptr_arith1_t_12345_r2");
         assert::i32eq(*p, 3);
@@ -27,24 +27,24 @@ ptr_arith1_t_12345_r2(p: i32*, n: i64): i32
 }
 
 export proc
-sum_array2_t12345_t5_r15(p: i32*, n: i64): i32
+sum_array2_t12345_t5_r15(p: i32*, n: size_t): i32
 {
         log::id("ptrs::sum2_array_t12345_t5_r15");
 
         let sum: i32 = 0;
-        for (let i: i64 = 0; i < n; i += 1) {
+        for (let i: size_t = 0; i < n; i += 1) {
                 sum += *(p+i);
         }
         return sum;
 }
 
 export proc
-sum_array1_t12345_t5_r15(p: i32*, n: i64): i32
+sum_array1_t12345_t5_r15(p: i32*, n: size_t): i32
 {
         log::id("ptrs::sum1_array_t12345_t5_r15");
 
         let sum: i32 = 0;
-        for (let i: i64 = 0; i < n; i += 1) {
+        for (let i: size_t = 0; i < n; i += 1) {
                 sum += p[i];
         }
         return sum;
